fix: modified cartitem.java and modified cart item test.
