+++ app/src/main/java/q4/team1/CartItem.java
@@ -18,6 +18,7 @@ public class CartItem {
     }
     this.orgPrice = orgPrice;
     this.price = orgPrice*quantity;
+    
   }
   public String getName(){
     return name;
